// packages/database/prisma-generated/zod/outputTypeSchemas/JrFindReplaceRejectFindManyArgsSchema.ts
import { z } from 'zod';
import type { Prisma } from '../../client';
import { JrFindReplaceRejectIncludeSchema } from '../inputTypeSchemas/JrFindReplaceRejectIncludeSchema'
import { JrFindReplaceRejectWhereInputSchema } from '../inputTypeSchemas/JrFindReplaceRejectWhereInputSchema'
import { JrFindReplaceRejectOrderByWithRelationInputSchema } from '../inputTypeSchemas/JrFindReplaceRejectOrderByWithRelationInputSchema'
import { JrFindReplaceRejectWhereUniqueInputSchema } from '../inputTypeSchemas/JrFindReplaceRejectWhereUniqueInputSchema'
import { JrFindReplaceRejectScalarFieldEnumSchema } from '../inputTypeSchemas/JrFindReplaceRejectScalarFieldEnumSchema'
import { BrandFindManyArgsSchema } from "../outputTypeSchemas/BrandFindManyArgsSchema"
import { LocationFindManyArgsSchema } from "../outputTypeSchemas/LocationFindManyArgsSchema"
import { TaxonomyFindManyArgsSchema } from "../outputTypeSchemas/TaxonomyFindManyArgsSchema"
import { StoryFindManyArgsSchema } from "../outputTypeSchemas/StoryFindManyArgsSchema"
import { FandomFindManyArgsSchema } from "../outputTypeSchemas/FandomFindManyArgsSchema"
import { SeriesFindManyArgsSchema } from "../outputTypeSchemas/SeriesFindManyArgsSchema"
import { CastFindManyArgsSchema } from "../outputTypeSchemas/CastFindManyArgsSchema"
import { JrExtractionRuleFindManyArgsSchema } from "../outputTypeSchemas/JrExtractionRuleFindManyArgsSchema"
import { JrFindReplaceRejectCountOutputTypeArgsSchema } from "../outputTypeSchemas/JrFindReplaceRejectCountOutputTypeArgsSchema"
// Select schema needs to be in file to prevent circular imports
//------------------------------------------------------

export const JrFindReplaceRejectSelectSchema: z.ZodType<Prisma.JrFindReplaceRejectSelect> = z.object({
  id: z.boolean().optional(),
  lookFor: z.boolean().optional(),
  replaceWith: z.boolean().optional(),
  ruleAction: z.boolean().optional(),
  isRegex: z.boolean().optional(),
  regexFlags: z.boolean().optional(),
  priority: z.boolean().optional(),
  brands: z.union([z.boolean(),z.lazy(() => BrandFindManyArgsSchema)]).optional(),
  locations: z.union([z.boolean(),z.lazy(() => LocationFindManyArgsSchema)]).optional(),
  taxonomies: z.union([z.boolean(),z.lazy(() => TaxonomyFindManyArgsSchema)]).optional(),
  stories: z.union([z.boolean(),z.lazy(() => StoryFindManyArgsSchema)]).optional(),
  fandoms: z.union([z.boolean(),z.lazy(() => FandomFindManyArgsSchema)]).optional(),
  series: z.union([z.boolean(),z.lazy(() => SeriesFindManyArgsSchema)]).optional(),
  casts: z.union([z.boolean(),z.lazy(() => CastFindManyArgsSchema)]).optional(),
  extractionRules: z.union([z.boolean(),z.lazy(() => JrExtractionRuleFindManyArgsSchema)]).optional(),
  _count: z.union([z.boolean(),z.lazy(() => JrFindReplaceRejectCountOutputTypeArgsSchema)]).optional(),
}).strict()

export const JrFindReplaceRejectFindManyArgsSchema: z.ZodType<Prisma.JrFindReplaceRejectFindManyArgs> = z.object({
  select: JrFindReplaceRejectSelectSchema.optional(),
  include: z.lazy(() => JrFindReplaceRejectIncludeSchema).optional(),
  where: JrFindReplaceRejectWhereInputSchema.optional(),
  orderBy: z.union([ JrFindReplaceRejectOrderByWithRelationInputSchema.array(),JrFindReplaceRejectOrderByWithRelationInputSchema ]).optional(),
  cursor: JrFindReplaceRejectWhereUniqueInputSchema.optional(),
  take: z.number().optional(),
  skip: z.number().optional(),
  distinct: z.union([ JrFindReplaceRejectScalarFieldEnumSchema,JrFindReplaceRejectScalarFieldEnumSchema.array() ]).optional(),
}).strict() ;

export default JrFindReplaceRejectFindManyArgsSchema;